+++ src/cms/preview-templates/IndexPagePreview.js
@@ -16,7 +16,6 @@ const IndexPagePreview = ({ entry, getAsset }) => {
         bottomSignIn={data.bottomSignIn}
         teachersSection={data.teachersSection}
         newsSection={data.newsSection}
-        partners={[]}
       />
     )
   } else {
